app/requirements.txt
#flask
Flask==0.11
aniso8601==1.1.0
click==6.6
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pytz==2016.4
six==1.10.0
Werkzeug==0.11.10

#flask extensions
Flask-RESTful==0.3.5
Flask-HTTPAuth==3.2.1
flask-cors==2.1.2
Flask-Session==0.3

#DB driver
pymongo==3.2.2

#Python goodies
requests==2.11.0
python-dateutil==2.5.3

#Crypto
passlib==1.6.5
python-jose==1.3.2

#wsgi container
uwsgi
